Move tests
Move tests in alivedetection_tests.c to
util_tests.c.
